+++ components/x-gift-article/src/GiftArticle.scss
@@ -12,7 +12,7 @@
 @import 'Message.scss';
 @import 'MobileShareButtons.scss';
 @import 'HighlightSection.scss';
-@import 'HighlightMessage.scss';
+@import 'HighlightRecipientMessage.scss';
 @import 'HighlightSavedMessage.scss';
