+++ components/x-gift-article/src/HighlightRecipientMessage.scss
